+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 _Changes in the next release_
 
+### Changed
+
+- Add `stable_ids` option to media browse and search requests.
+
 ---
 
 ## 0.4.0-beta - 2026-03-16